modified2

+++ src/BmiService.java
@@ -1,6 +1,6 @@
 public class BmiService {
-    public float calculate(float weight, float height) {
+    public int calculate(float weight, float height) {
         float bmi = weight / (height * height);
-        return bmi;
+        return (int) bmi;
     }
 }
